externalize ObservationValidationInfo
